+++ tools/release/lib/utils.js
@@ -1,12 +1,12 @@
 /**
  * External dependencies
  */
-// @ts-ignore
 const fs = require( 'fs' );
 const childProcess = require( 'child_process' );
 const { randomUUID } = require( 'crypto' );
 const path = require( 'path' );
 const os = require( 'os' );
+// @ts-ignore
 const { confirm } = require( '@inquirer/prompts' );
 
 /**